resolve bug and program works
