--- gustavo_ygarzabal/topic2/src/main/java/Building.java
@@ -1,5 +1,0 @@
-public interface Building {
-    void work();
-    void showProducts();
-    void cleanProducts();
-}
