+++ gustavo_ygarzabal/topic2/src/main/java/Main.java
@@ -1,5 +1,8 @@
 import com.globant.bootcamp.abstracts.Animal;
+import com.globant.bootcamp.buildings.Building;
 import com.globant.bootcamp.buildings.Farm;
+import com.globant.bootcamp.buildings.HenHouse;
+import com.globant.bootcamp.buildings.HenHouseCreator;
 import com.globant.bootcamp.enums.Gender;
 import com.globant.bootcamp.roles.*;
 
@@ -9,15 +12,17 @@ public class Main {
     public static void main(String[] args) {
         Farm farm = Farm.getInstance();
 
-
-        farm.setDemoConfig(5, new Farmer(Gender.MALE, "Bob"));
-
        //Method already implemented and use, i mantain it. It could be deleted
         Animal[] animals = new Animal[farm.getAnimals().size()];
         farm.getAnimals().toArray(animals);
-        Main.singAnimals(animals);
+        singAnimals(animals);
 
 
+        HenHouse henHouse = (HenHouse) new HenHouseCreator().createBuilding((int)(40*0.7),(int)(40*0.3));
+        ArrayList<Building> buildings = new ArrayList<>();
+        buildings.add(henHouse);
+
+        farm.setBuildings(buildings);
         farm.work();
         farm.getProducts().forEach(product -> product.printProduct());
     }
