+++ gustavo_ygarzabal/topic2/src/main/java/com/globant/bootcamp/buildings/AnimalManager.java
@@ -1,3 +1,5 @@
+package com.globant.bootcamp.buildings;
+
 import com.globant.bootcamp.abstracts.Animal;
 
 import java.util.ArrayList;
