+++ gustavo_ygarzabal/topic2/src/main/java/com/globant/bootcamp/buildings/BuildingCreator.java
@@ -1,3 +1,5 @@
+package com.globant.bootcamp.buildings;
+
 public abstract class BuildingCreator {
     public abstract Building createBuilding(int capacity);
 }
