+++ gustavo_ygarzabal/topic2/src/main/java/com/globant/bootcamp/buildings/FarmCreator.java
@@ -1,3 +1,5 @@
+package com.globant.bootcamp.buildings;
+
 public class FarmCreator extends BuildingCreator{
     @Override
     public Building createBuilding(int capacity) {
