+++ gustavo_ygarzabal/topic2/src/main/java/com/globant/bootcamp/buildings/HenHouseCreator.java
@@ -1,3 +1,6 @@
+package com.globant.bootcamp.buildings;
+
+import com.globant.bootcamp.AnimalsFactory.ChickenCreator;
 import com.globant.bootcamp.animals.Chicken;
 import com.globant.bootcamp.enums.EggColor;
 
@@ -11,8 +14,8 @@ public class HenHouseCreator extends BuildingCreator{
 
     public Building createBuilding(int red, int white) {
         HenHouse henHouse = (HenHouse) createBuilding(red+white);
-        henHouse.addAllowedAnimal((Chicken)ChickenCreator.getHen(EggColor.RED));
-        henHouse.addAllowedAnimal((Chicken)ChickenCreator.getHen(EggColor.WHITE));
+        henHouse.addAllowedAnimal((Chicken) ChickenCreator.getHen(EggColor.RED));
+        henHouse.addAllowedAnimal((Chicken) ChickenCreator.getHen(EggColor.WHITE));
         henHouse.getAnimals().addAll(ChickenCreator.getHen(EggColor.RED,red));
         henHouse.getAnimals().addAll(ChickenCreator.getHen(EggColor.WHITE,white));
         return henHouse;
